// frontend/src/shared/types/receipt.ts
export interface MerchantReceipt {
    owner: string;
    receiptHash: string;
    invoiceHash: string;
    amount: number;
    tokenType: number;
    merchantNote: string;
    timestamp?: number;
    created_at?: string;
    transactionId?: string;
}